psy/requirements.txt
altgraph==0.17.3
anyio==3.7.1
asgiref==3.6.0
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==41.0.2
defusedxml==0.7.1
Django==4.2
django-allauth==0.54.0
django-datetime-widget==0.9.3
django-debug-toolbar==4.1.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.18
et-xmlfile==1.1.0
google-api-core==2.11.1
google-api-python-client==2.94.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
idna==3.4
numpy==1.25.1
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
pandas==2.0.3
pefile==2023.2.7
Pillow==9.5.0
protobuf==4.23.4
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.5
PyJWT==2.7.0
pyparsing==3.1.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyTelegramBotAPI==4.12.0
python-dateutil==2.8.2
python-telegram-bot==20.4
python3-openid==3.2.0
pytz==2023.3
pywin32-ctypes==0.2.2
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
telebot==0.0.5
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.13
